Handle redirects in the inline pagelet processor in simple cases.
This does not cope with detecting redirect-loops, nor with cases
in which we shouldn't really be using the inline method (i.e.
redirecting to pages not on the same host).

Modified features slightly so some text is shown for XHR requests
instead of being hidden - this makes testing with multiple
pagelets in the same page easier.